Organize v1 routes in separate directories
